+++ api/v3/Campagnodon/Start.php
@@ -166,7 +166,11 @@ function _civicrm_api3_campagnodon_Start_spec(&$spec) {
  */
 function civicrm_api3_campagnodon_Start($params) {
   $tx = new CRM_Core_Transaction();
+  // CiviRules is buggy is we only have 'Campagnodon api' permission... So we are cheating by changing the permissions during the API call.
+  $previous_permissions = CRM_Core_Config::singleton()->userPermissionClass->permissions;
   try {
+    CRM_Core_Config::singleton()->userPermissionClass->permissions = ['access CiviCRM', 'access Campagnodon', 'administer CiviCRM', 'Campagnodon api'];
+
     // checking API version
     if ($params['campagnodon_version'] !== '1') {
       throw new API_Exception("Unkwnown API version '".($params['campagnodon_version'] ?? '')."'");
@@ -361,6 +365,8 @@ function civicrm_api3_campagnodon_Start($params) {
   } catch (Throwable $e) {
     $tx->rollback();
     throw $e;
+  } finally {
+    CRM_Core_Config::singleton()->userPermissionClass->permissions = $previous_permissions;
   }
   $tx->commit();